+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Unreleased
 
+- Minor: Include loot source category in notification metadata. (#196)
 - Minor: Avoid death notifications for more safe activities: Barbarian Assault, Chambers of Xeric, Clan Wars, Last Man Standing, Nightmare Zone, Soul Wars, TzHaar Fight Pit. (#194)
 - Minor: Avoid unwarranted settings warnings when world hopping. (#183)
 - Minor: Include number of quest completions and quest points in quest notifications. (#178)